@arnaud-ma 2024-08-29
Ah, the coverage diff  is because the `get_subsequent_assign_lines` and `source_line_to_tokens` functions were not explicitly tested before. Now that they are no longer used in the code, it only shows up here